refactor: moved PushX methods from BufferUtils to a new namespace
